+++ Changelog.md
@@ -32,6 +32,7 @@
 
 ## Other changes:
 - [Add stonecutting recipes for chisel and factory blocks - by @redcatone](https://github.com/TheStaticVoid/StaTech-Industry/pull/402)
+- [Added new processing chain to make chlorine and sodium hydroxide from salt water](https://github.com/TheStaticVoid/StaTech-Industry/issues/412)
 - [Add GregSchizo to Emojiful](https://github.com/TheStaticVoid/StaTech-Industry/issues/396)
 - [Fixed Crimson Berries being consumed in Photosynthetic Chamber](https://github.com/TheStaticVoid/StaTech-Industry/issues/396)
 - [Set Faster Texture Stitching to false](https://github.com/TheStaticVoid/StaTech-Industry/issues/398) - Addresses the weird texture bugs when Iris is not installed